Fix Activiti bug where TIMESTAMP columns used a default of "0000-00-00 00:00:00"
The activiti engineers "fixed" this in 5.12, but they put the fix in the 5.11
upgrade script after 5.11 was released.  To work around this, we are applying
the fix to a fork of the 5.11 code and we'll release our own whaleshark 5.11
branch.  See http://forums.activiti.org/en/viewtopic.php?f=6&t=5707 for
additional background.